refactor: Пакет приема сообщений вынесен из client
