+++ app/deps/client.py
@@ -1,5 +1,5 @@
 from services.client import MatrixClient
-from services.client.receive.rgb import RGBMessageReceiver
+from services.receive.rgb import RGBMessageReceiver
 from services.client.stream import StreamMatrixClient
 from services.stream.ws import WsStream
 from starlette.websockets import WebSocket
