--- app/services/client/receive/__init__.py
@@ -1,3 +1,0 @@
-from services.client.receive.proto import MessageReceiver
-
-__all__ = ['MessageReceiver']
