+++ app/services/receive/__init__.py
@@ -1,0 +1,3 @@
+from services.receive.proto import MessageReceiver
+
+__all__ = ['MessageReceiver']
