+++ app/services/receive/rgb.py
@@ -5,7 +5,7 @@ from dto.pixel import RGBPixel
 from pydantic import TypeAdapter
 from pydantic import ValidationError
 from services.client.data import Message
-from services.client.receive.proto import MessageReceiver
+from services.receive.proto import MessageReceiver
 from services.stream import Stream
 
 logger = logging.getLogger(__name__)
