+++ app/services/stream/ws.py
@@ -21,7 +21,7 @@ class WsStream(Stream):
 
     async def wait_json(self) -> Any:
         """
-        Ожидание данных от клиента, для передачи на матрицу
+        Ожидание данных от клиента
         :raises:
             ValueError: В случае проблем с обработкой входных данных
         """
